requirements.txt
# E-Raksha Main Requirements
torch==2.1.0
torchvision==0.16.0
torchaudio==2.1.0
numpy==1.24.3
opencv-python-headless==4.8.1.78
Pillow==10.1.0
scikit-learn==1.3.2
matplotlib==3.7.2
seaborn==0.12.2
tqdm==4.66.1
requests==2.31.0
python-dotenv==1.0.0